se cambia propiedad numero de orden a Long se personaliza mensage en propertis

+++ src/main/java/mx/gob/renapo/registrocivil/actos/defuncion/dto/ActaDefuncionDTO.java
@@ -37,7 +37,7 @@ public class ActaDefuncionDTO {
     @Autowired
     private CatDestinoCadaverDTO destinoCadaver;
     private String domicilioPanteon;
-    private Integer noOrden;
+    private Long noOrden;
 
     private Integer mesesNacimientoFinado;
     private Date horaNacimientoFinado;
